refactor: split hooks
