+++ packages/shared/widget-plugin-external-events/src/hooks/useOnResetFiltersEvent.ts
@@ -1,0 +1,13 @@
+import { useRef } from "react";
+import { useTransmit } from "./useTransmit";
+import { $events } from "../common";
+
+export function useOnResetFiltersEvent(inputChannel: string, broadcastChannel: string): void {
+    useTransmit(
+        useRef<Parameters<typeof useTransmit>[0]>({
+            inputChannel,
+            event: $events.reset.filters,
+            onEvent: emit => emit(broadcastChannel, $events.reset.value)
+        }).current
+    );
+}
